Split unit tests
